Separate PayoutDetailsForm to PayoutDetailsFormIndividual and PayoutDetailsFormCompany
